pyproject.toml
[tool.poetry]
name = "crypto-sentiment-analysis"
version = "0.1.0"
description = "Airflow DAGs for crypto sentiment analysis"
authors = ["Andrew Bakos <andrew@romutrechtregion.nl>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
apache-airflow = "^2.9.0"
apache-airflow-providers-google = "^10.11.1"
dbt-bigquery = "^1.9.1"
streamlit = "^1.42.1"
google-cloud-bigquery = "^3.29.0"
google-cloud-storage = "^2.16.0"
requests = "^2.32.3"
pandas = "^2.2.3"
textblob = "^0.19.0"
nltk = "^3.8.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
